400. Async Await

+++ controllers/feed.js
@@ -1,14 +1,26 @@
 const fs = require('fs')
 const path = require('path')
 const { validationResult } = require('express-validator/check')
+
 const Post = require('../models/post')
+const User = require('../models/user')
 
 exports.getPosts = (req, res, next) => {
-    Post.find()
+    const currentPage = req.query.page || 1
+    const perPage = 2
+    let totalItems
+    Post.find().countDocuments()
+        .then(count => {
+            totalItems = count
+            return Post.find()
+                .skip((currentPage - 1) * perPage)
+                .limit(perPage)
+        })
         .then(posts => {
             res.status(200).json({
                 message: 'Fetched posts successfully.',
-                posts: posts
+                posts: posts,
+                totalItems: totalItems
             })
         })
         .catch(err => {
@@ -35,17 +47,30 @@ exports.createPost = (req, res, next) => {
     console.log(imageUrl)
     const title     = req.body.title
     const content   = req.body.content
+    let creator
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {name: 'Ardiyan Agus'}
+        creator: req.userId
     })
     post.save()
         .then(result => {
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user
+            user.posts.push(post)
+            return user.save()
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully',
-                post: result
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             })
         })
         .catch(err => {
@@ -89,7 +114,6 @@ exports.updatePost = (req, res, next) => {
     const title = req.body.title
     const content = req.body.content
     let imageUrl = req.body.image
-    console.log(imageUrl)
     if(req.file) {
         imageUrl = req.file.path.replace("\\","/")
     }
@@ -103,6 +127,11 @@ exports.updatePost = (req, res, next) => {
             if(!post){
                 const error = new Error('Could not find post.')
                 error.statusCode = 404
+                throw error
+            }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized')
+                error.statusCode = 403
                 throw error
             }
             if(imageUrl !== post.imageUrl){
@@ -136,13 +165,44 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404
                 throw error
             }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized')
+                error.statusCode = 403
+                throw error
+            }
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId)
         })
         .then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save()
+        })
+        .then(result => {
             res.status(200).json({
                 message: 'Deleted Post.'
+            })
+        })
+        .catch(err => {
+            if(!err.statusCode){
+                err.statusCode = 500
+            }
+            next(err)
+        })
+}
+
+exports.status = (req, res, next) => {
+    User.findById(req.userId)
+        .then(user => {
+            if(!user){
+                const error = new Error('User not found')
+                error.statusCode = 404
+                throw error
+            }
+            res.status(200).json({
+                status: user.status
             })
         })
         .catch(err => {
